Non-Volatile Blacklist
Forced a directory create for saving the data. Set up the functions to save and read the Blacklist from storage. Forced the cmake compiler to make debugging logs (???) as to use pahole to see paddings in memory

+++ fast_path/ips_fast_common.h
@@ -19,7 +19,8 @@ struct ips_token_bucket {
 
 struct ips_blocklist_data {
     __u64 ban_timestamp;
-    __u8 is_static;      // 0 = Auto-banned (Dynamic), 1 = Threat Intel (Permanent)
+    __u64 is_static;      // 0 = Auto-banned (Dynamic), 1 = Threat Intel (Permanent)
+    // is_static was made __64 from __8 to avoid padding
 };
 
 #endif /* IPS_FAST_COMMON_H */
